+++ app/src/main/java/com/example/androidtrainingproject/InternetConnection.kt
@@ -1,0 +1,42 @@
+package com.example.androidtrainingproject
+
+import android.annotation.SuppressLint
+import android.content.Context
+import android.net.ConnectivityManager
+import android.net.NetworkCapabilities
+import android.os.Build
+
+class InternetConnection() {
+
+    //@SuppressLint("ObsoleteSdkInt")
+    private fun isConnected(context: Context): Boolean {
+        val connectivityManager =
+            this.getSystemService(Context.CONNECTIVITY_SERVICE) as ConnectivityManager
+
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.M) {
+            val network = connectivityManager.activeNetwork ?: return false
+            val activeNetwork = connectivityManager.getNetworkCapabilities(network) ?: return false
+
+            return when {
+
+                activeNetwork.hasTransport(NetworkCapabilities.TRANSPORT_WIFI) -> {
+                    true
+                }
+                activeNetwork.hasTransport(NetworkCapabilities.TRANSPORT_CELLULAR) -> {
+                    true
+                }
+
+                else -> {
+                    false
+                }
+            }
+        } else {
+
+            @Suppress("DEPRECATION") val networkInfo =
+                connectivityManager.activeNetworkInfo ?: return false
+            @Suppress("DEPRECATION") return networkInfo.isConnected
+        }
+
+    }
+
+}
